// src/redux/types/Pokemon.ts
export interface PokemonQuery {
    name: string, 
    url: string
}